Moved Cell to its own file
There’s no reason for cell to be with Grid if Player, Entity, and Fence
are their own files.
